# src/main/java/ar/edu/utn/frc/tup/lciii/dtos/common/CountryRequestDTO.java
package ar.edu.utn.frc.tup.lciii.dtos.common;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

@Builder
@Data
@NoArgsConstructor
@AllArgsConstructor
public class CountryRequestDTO {
    private int cantPaises;
}