# src/main/java/ar/edu/utn/frc/tup/lciii/entity/CountryEntity.java
package ar.edu.utn.frc.tup.lciii.entity;

import jakarta.persistence.Column;
import jakarta.persistence.Entity;
import jakarta.persistence.Id;
import jakarta.persistence.Table;
import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@Entity
@Table(name = "country")
@Data
@AllArgsConstructor
@NoArgsConstructor
public class CountryEntity {
    @Id
    @Column
    private String code;

    @Column
    private String name;
}